requirements.txt
tensorflow 
streamlit 
numpy
pandas
pillow
opencv_python
scikit_learn
